Added M1-M4 Button prompts to Shared
To account for Apple MFi Controllers and the soon-to-be-released HORIPAD for Steam: the M1-M4 buttons are added. They're equivalent to the Back Buttons found in Xbox Elite, DualSense Edge and Steam Hardware Inputs (Steam Controller and Steam Deck)

If you're shipping a game on Mobile and you wanna support the Back Buttons based on the Operating System's Gamepad APIs; please use the M1-M4 button prompts for Generic Controllers only.